Incomplete, committing anyway, see details
This was incomplete when I wrote it, and since I'm not working on it at
the moment, I'm committing the incomplete version. I don't have time right
now to go through it, but I should still push it, since it's a bad idea to
have something uncommitted sitting on my drive for so long.